Implemented basic Event/Guild Repository test coverage and fixed discovered bugs
Test coverage for EventRepository and GuildRepository was left around 60%. The intention is to return to these tests and investigate effectively mocking the DBAL query builder which would allow for more effective repository testing while eliminating the need for a database connection.
A few type bugs were discovered that needed to be corrected on the Repository interface. The intention was to support other implementation such as UUIDv4 primary keys, so allowing both strings and integers was preferred over typecasting where appropriate.